完成：商品编辑

+++ src/main/java/com/dyl/o2o/common/util/ImageUtil.java
@@ -3,6 +3,7 @@ package com.dyl.o2o.common.util;
 import lombok.extern.slf4j.Slf4j;
 import net.coobird.thumbnailator.Thumbnails;
 import net.coobird.thumbnailator.geometry.Positions;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.imageio.ImageIO;
@@ -11,6 +12,7 @@ import java.io.File;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.List;
 import java.util.Random;
 
 /**
@@ -85,22 +87,36 @@ public class ImageUtil {
 
 
     /**
-     * 根据用户上传的文件生成带水印的缩略图
+     * 根据用户上传的文件生成带水印的缩略图，保存到服务器
      * @param newImg
      * @return
      * @throws IOException
      */
     public static String uploadThumbnail(MultipartFile newImg) throws IOException {
+        //入参校验
+        if (ObjectUtils.isEmpty(newImg.getOriginalFilename())){
+            return null;
+        }
         String imgFileAbsolutePath = uploadFile(newImg);
-
+        File img = new File(imgFileAbsolutePath);
         //用服务器上的图片生成带水印的缩略图，并将原图覆盖
         ImageUtil.img2Thumbnail(img);
         return imgFileAbsolutePath;
     }
 
+    /**
+     * 保存用户上传图片到服务器
+     * @param userFile
+     * @return
+     * @throws IOException
+     */
     public static String uploadFile(MultipartFile userFile) throws IOException {
+        //入参校验
+        if (ObjectUtils.isEmpty(userFile.getOriginalFilename())){
+            return null;
+        }
         //生成服务器上保存图片的地址
-        String fileAbsolutePath = PathUtil.getImgBasePath() +"\\"+ ImageUtil.getRandomFileName() + ImageUtil.getFileNameExtension(newImg.getOriginalFilename());
+        String fileAbsolutePath = PathUtil.getImgBasePath() +"\\"+ ImageUtil.getRandomFileName() + ImageUtil.getFileNameExtension(userFile.getOriginalFilename());
         //根据地址创建文件
         File file = new File(fileAbsolutePath);
         //若文件地址不存在，生成地址
@@ -114,17 +130,29 @@ public class ImageUtil {
     }
 
     /**
-     * 删除文件
-     * @param shopImg
+     * 批量删除文件
+     * @param filePathList
      */
-    public static void deleteImg(String shopImg) {
-        File file = new File(shopImg);
+    public static void deleteFileList(List<String> filePathList){
+        for (String path : filePathList) {
+            deleteFile(path);
+        }
+    }
+
+    /**
+     * 删除文件
+     * @param filePath
+     */
+    public static void deleteFile(String filePath) {
+        //入参校验
+        if (ObjectUtils.isEmpty(filePath)){
+            return;
+        }
+        File file = new File(filePath);
         if (file.exists()){
             file.delete();
         }
     }
-
-
 
     /**
      * 创建一个带有水印的缩略图
